+++ agixt/Memories.py
@@ -511,186 +511,3 @@ class Memories:
         # Sort the chunks by their score in descending order before returning them
         content_chunks.sort(key=lambda x: x[0], reverse=True)
         return [chunk_text for score, chunk_text in content_chunks]
-
-    async def get_context(
-        self,
-        user_input: str,
-        limit: int = 10,
-        websearch: bool = False,
-        additional_collections: List[str] = [],
-    ) -> str:
-        self.collection_number = 0
-        context = await self.get_memories(
-            user_input=user_input,
-            limit=limit,
-            min_relevance_score=0.2,
-        )
-        self.collection_number = 2
-        positive_feedback = await self.get_memories(
-            user_input=user_input,
-            limit=3,
-            min_relevance_score=0.7,
-        )
-        self.collection_number = 3
-        negative_feedback = await self.get_memories(
-            user_input=user_input,
-            limit=3,
-            min_relevance_score=0.7,
-        )
-        if positive_feedback or negative_feedback:
-            context.append(
-                f"The users input makes you to remember some feedback from previous interactions:\n"
-            )
-            if positive_feedback:
-                context += f"Positive Feedback:\n{positive_feedback}\n"
-            if negative_feedback:
-                context += f"Negative Feedback:\n{negative_feedback}\n"
-        if websearch:
-            self.collection_number = 1
-            context += await self.get_memories(
-                user_input=user_input,
-                limit=limit,
-                min_relevance_score=0.2,
-            )
-        if additional_collections:
-            for collection in additional_collections:
-                self.collection_number = collection
-                context += await self.get_memories(
-                    user_input=user_input,
-                    limit=limit,
-                    min_relevance_score=0.2,
-                )
-        return context
-
-    async def batch_prompt(
-        self,
-        user_inputs: List[str] = [],
-        prompt_name: str = "Ask Questions",
-        prompt_category: str = "Default",
-        batch_size: int = 10,
-        **kwargs,
-    ):
-        i = 0
-        tasks = []
-        responses = []
-        if user_inputs == []:
-            return []
-        for user_input in user_inputs:
-            i += 1
-            logging.info(f"[{i}/{len(user_inputs)}] Running Prompt: {prompt_name}")
-            if i % batch_size == 0:
-                responses += await asyncio.gather(**tasks)
-                tasks = []
-            task = asyncio.create_task(
-                await self.ApiClient.prompt_agent(
-                    agent_name=self.agent_name,
-                    prompt_name=prompt_name,
-                    prompt_args={
-                        "prompt_category": prompt_category,
-                        "user_input": user_input,
-                        **kwargs,
-                    },
-                )
-            )
-            tasks.append(task)
-        responses += await asyncio.gather(**tasks)
-        return responses
-
-    # Answer a question with context injected, return in sharegpt format
-    async def agent_dpo_qa(self, question: str = "", context_results: int = 10):
-        context = await self.get_context(user_input=question, limit=context_results)
-        prompt = f"### Context\n{context}\n### Question\n{question}"
-        chosen = await self.ApiClient.prompt_agent(
-            agent_name=self.agent_name,
-            prompt_name="Answer Question with Memory",
-            prompt_args={
-                "prompt_category": "Default",
-                "user_input": question,
-                "context_results": context_results,
-            },
-        )
-        # Create a memory with question and answer
-        self.collection_number = 0
-        await self.write_text_to_memory(
-            user_input=question,
-            text=chosen,
-            external_source="Synthetic QA",
-        )
-        rejected = await self.ApiClient.prompt_agent(
-            agent_name=self.agent_name,
-            prompt_name="Wrong Answers Only",
-            prompt_args={
-                "prompt_category": "Default",
-                "user_input": question,
-            },
-        )
-        return prompt, chosen, rejected
-
-    # Creates a synthetic dataset from memories in sharegpt format
-    async def create_dataset_from_memories(
-        self, dataset_name: str = "", batch_size: int = 10
-    ):
-        self.agent_config["settings"]["training"] = True
-        self.ApiClient.update_agent_settings(
-            agent_name=self.agent_name, settings=self.agent_config["settings"]
-        )
-        memories = []
-        questions = []
-        if dataset_name == "":
-            dataset_name = f"{datetime.now().isoformat()}-dataset"
-        collections = await self.get_collections()
-        for collection in collections:
-            self.collection_name = collection
-            memories += await self.export_collection_to_json()
-        logging.info(f"There are {len(memories)} memories.")
-        memories = [memory["text"] for memory in memories]
-        # Get a list of questions about each memory
-        question_list = self.batch_prompt(
-            user_inputs=memories,
-            batch_size=batch_size,
-        )
-        for question in question_list:
-            # Convert the response to a list of questions
-            question = question.split("\n")
-            question = [
-                item.lstrip("0123456789.*- ") for item in question if item.lstrip()
-            ]
-            question = [item for item in question if item]
-            question = [item.lstrip("0123456789.*- ") for item in question]
-            questions += question
-        prompts = []
-        good_answers = []
-        bad_answers = []
-        for question in questions:
-            prompt, chosen, rejected = await self.agent_dpo_qa(
-                question=question, context_results=10
-            )
-            prompts.append(prompt)
-            good_answers.append(
-                [
-                    {"content": prompt, "role": "user"},
-                    {"content": chosen, "role": "assistant"},
-                ]
-            )
-            bad_answers.append(
-                [
-                    {"content": prompt, "role": "user"},
-                    {"content": rejected, "role": "assistant"},
-                ]
-            )
-        dpo_dataset = {
-            "prompt": questions,
-            "chosen": good_answers,
-            "rejected": bad_answers,
-        }
-        # Save messages to a json file to be used as a dataset
-        os.makedirs(f"./WORKSPACE/{self.agent_name}/datasets", exist_ok=True)
-        with open(
-            f"./WORKSPACE/{self.agent_name}/datasets/{dataset_name}.json", "w"
-        ) as f:
-            f.write(json.dumps(dpo_dataset))
-        self.agent_config["settings"]["training"] = False
-        self.ApiClient.update_agent_settings(
-            agent_name=self.agent_name, settings=self.agent_config["settings"]
-        )
-        return dpo_dataset
